fix(flpdf): u64 range bound for outline group_length (gemini review, flpdf-rm09)
Outlines Hint Table group_length iterated (first_object..first_object+nobjects)
with u32 arithmetic on the upper bound, which could panic in debug builds on a
pathological (>4-billion-object) layout. Widen the range bound to u64 and cast
back to u32 for the byte_lengths lookup. Byte-neutral for any realistic document
(values fit in u32); all goldens unchanged. Aligns with review-pattern #3
(verify external/derived integers before unsigned arithmetic).

+++ crates/flpdf/src/linearization/writer.rs
@@ -2889,8 +2889,13 @@ pub fn write_linearized<R: Read + Seek>(
                         ))
                     })? as u32;
                 // cov:ignore-end
-                let group_length: u64 = (info.first_object..info.first_object + info.nobjects)
-                    .map(|n| byte_lengths.get(&n).copied().unwrap_or(0) as u64)
+                // `u64` range bound prevents a `u32` overflow panic on a
+                // pathological (>4-billion-object) layout; for any realistic
+                // document the values fit in `u32`, so `n as u32` and the sum
+                // are byte-identical to the direct computation.
+                let group_length: u64 = (info.first_object as u64
+                    ..info.first_object as u64 + info.nobjects as u64)
+                    .map(|n| byte_lengths.get(&(n as u32)).copied().unwrap_or(0) as u64)
                     .sum();
                 Ok(OutlineHintTable {
                     first_object: info.first_object,
